feat: rename RadioGroup Component
