+++ packages/components/src/components/index.ts
@@ -15,6 +15,7 @@ export { default as DropDown } from './atoms/form/DropDown'
 export { default as InfoTooltip } from './atoms/form/InfoTooltip'
 export { default as InputTextWithoutError } from './atoms/form/InputTextWithoutError'
 export { default as InputText } from './atoms/form/InputText'
+export { default as RadioGroupWithoutError } from './atoms/form/RadioGroupWithoutError'
 export { default as RadioGroup } from './atoms/form/RadioGroup'
 export { default as Range } from './atoms/form/Range'
 export { default as RangeError } from './atoms/form/RangeError'
